requirements.txt
jedi==0.19.1
attrs==24.2.0
requests==2.28.1
filelock==3.13.1
python-dotenv==1.0.1
click==8.1.7
decorator==5.1.1
torchvision==0.16.0+cu118
aiosignal==1.3.1
pexpect==4.9.0
frozenlist==1.5.0
wcwidth==0.2.13
PyYAML==6.0.2
torch-geometric==2.6.1
GitPython==3.1.43
charset-normalizer==2.1.1
torch-spline-conv==1.2.2+pt21cu118
sympy==1.13.1
matplotlib-inline==0.1.7
asttokens==2.4.1
torch-sparse==0.6.18+pt21cu118
urllib3==1.26.13
pure_eval==0.2.3
networkx==3.2.1
executing==2.1.0
pip==22.0.2
torchaudio==2.1.0+cu118
typing_extensions==4.9.0
propcache==0.2.0
exceptiongroup==1.2.2
idna==3.4
docker-pycreds==0.4.0
stack-data==0.6.3
sentry-sdk==2.18.0
numpy==1.26.3
tqdm==4.67.0
aiohttp==3.10.10
traitlets==5.14.3
smmap==5.0.1
psutil==6.1.0
prompt_toolkit==3.0.48
scipy==1.14.1
torch-cluster==1.6.3+pt21cu118
wandb==0.18.6
torch==2.1.0+cu118
aiohappyeyeballs==2.4.3
Jinja2==3.1.3
Pygments==2.18.0
certifi==2022.12.7
six==1.16.0
multidict==6.1.0
pyparsing==3.2.0
ptyprocess==0.7.0
MarkupSafe==2.1.5
mpmath==1.3.0
platformdirs==4.3.6
async-timeout==4.0.3
setproctitle==1.3.3
yarl==1.17.1
ipython==8.29.0
setuptools==59.6.0
parso==0.8.4
gitdb==4.0.11
pillow==10.2.0
pyg-lib==0.4.0+pt21cu118
triton==2.1.0
fsspec==2024.2.0
protobuf==5.28.3
torch-scatter==2.1.2+pt21cu118
rootutils==1.0.7
